+++ src/Components/Authentication/Presentation/Form/LoginFormType.php
@@ -1,0 +1,22 @@
+<?php
+
+// src/Form/LoginFormType.php
+
+namespace App\Components\Authentication\Presentation\Form;
+
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\EmailType;
+use Symfony\Component\Form\Extension\Core\Type\PasswordType;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\Form\FormBuilderInterface;
+
+class LoginFormType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $builder, array $options) : void
+    {
+        $builder
+            ->add('email', EmailType::class)
+            ->add('password', PasswordType::class)
+            ->add('submit', SubmitType::class, ['label' => 'Login']);
+    }
+}
