+++ src/Components/Orderflow/Business/OrderFlowBusinessFacadeInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Components\Orderflow\Business;
+
+interface OrderFlowBusinessFacadeInterface
+{
+
+}
