+++ src/Components/Orderflow/Persistence/OrdersRepository.php
@@ -1,0 +1,65 @@
+<?php
+
+namespace App\Components\Orderflow\Persistence;
+
+use App\Entity\Orders;
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\Persistence\ManagerRegistry;
+
+/**
+ * @extends ServiceEntityRepository<Orders>
+ *
+ * @method Orders|null find($id, $lockMode = null, $lockVersion = null)
+ * @method Orders|null findOneBy(array $criteria, array $orderBy = null)
+ * @method Orders[]    findAll()
+ * @method Orders[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ */
+class OrdersRepository extends ServiceEntityRepository
+{
+    public function __construct(ManagerRegistry $registry)
+    {
+        parent::__construct($registry, Orders::class);
+    }
+
+    /**
+     * Find the most recent order by user's email.
+     *
+     * @param string $email The email of the user.
+     * @return Orders|null The most recent order or null if none found.
+     */
+    public function findMostRecentOrderByEmail(string $email): ?Orders
+    {
+        return $this->createQueryBuilder('o')
+            ->where('o.email = :email')
+            ->setParameter('email', $email)
+            ->orderBy('o.id', 'DESC')
+            ->setMaxResults(1)
+            ->getQuery()
+            ->getOneOrNullResult();
+    }
+
+//    /**
+//     * @return Orders[] Returns an array of Orders objects
+//     */
+//    public function findByExampleField($value): array
+//    {
+//        return $this->createQueryBuilder('o')
+//            ->andWhere('o.exampleField = :val')
+//            ->setParameter('val', $value)
+//            ->orderBy('o.id', 'ASC')
+//            ->setMaxResults(10)
+//            ->getQuery()
+//            ->getResult()
+//        ;
+//    }
+
+//    public function findOneBySomeField($value): ?Orders
+//    {
+//        return $this->createQueryBuilder('o')
+//            ->andWhere('o.exampleField = :val')
+//            ->setParameter('val', $value)
+//            ->getQuery()
+//            ->getOneOrNullResult()
+//        ;
+//    }
+}
