+++ src/Components/Registration/Business/RegistrationBusinessFacadeInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Components\Registration\Business;
+
+interface RegistrationBusinessFacadeInterface
+{
+
+}
