+++ src/Components/Registration/Persistence/UserEntityManager.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace App\Components\Registration\Persistence;
+
+use App\Entity\User;
+use App\Global\Persistence\DTO\UserDTO;
+use App\Global\Persistence\Mapping\Mapper;
+use Doctrine\ORM\EntityManagerInterface;
+
+class UserEntityManager
+{
+    public function __construct(
+        private EntityManagerInterface $entityManager,
+        private Mapper $mapper,
+    )
+    {}
+
+    public function register(UserDTO $userDTO): void
+    {
+        $user = $this->mapper->getUserEntity($userDTO);
+        $user->setPassword($this->hashPassword($user));
+        $this->entityManager->persist($user);
+        $this->entityManager->flush();
+    }
+    private function hashPassword(User $user) : string
+    {
+       return password_hash($user->getPassword(), PASSWORD_DEFAULT);
+    }
+}
