+++ src/Global/Persistence/DTO/UserDTO.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Global\Persistence\DTO;
+
+class UserDTO
+{
+
+    public string $username = '';
+    public string $email = '';
+    public string $password = '';
+}
