--- src/Model/DTO/ItemDTO.php
@@ -1,22 +1,0 @@
-<?php
-
-namespace App\Model\DTO;
-
-class ItemDTO
-{
-    public string $id = '';
-    public float $price = 0.00;
-    public string $name = '';
-    public string $nationality = '';
-    public string $position = '';
-    public string $thumbnail = '';
-    public string $clubName = '';
-    public string $clubAddress = '';
-    public string $clubFounded = '';
-    public string $clubWebsite = '';
-    public string $clubEmblem = '';
-    public int $team_id = 0;
-    public int $quantity = 0;
-    public string $user_id = '';
-    public int $item_id = 0;
-}
