--- src/Model/DTO/UserDTO.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Model\DTO;
-
-class UserDTO
-{
-
-    public string $username = '';
-    public string $email = '';
-    public string $password = '';
-}
